Added results checking
It was important to make sure each test is actually getting the
same results as the other libraries otherwise we aren't testing
the same thing and the results are INVALID.

Several issues came up while doing this.

1) CanvasMatrix multiplies on the right where as all the other
libraries multiply on the left. In other words, in other
libraries you do this

  m.identity();
  m.translate();
  m.rotateX();
  m.rotateY();
  m.scale();

Where as in CanvasMatrix you'd have to do this

  m.identity();
  m.scale();
  m.rotateY();
  m.rotateX();
  m.translate();

That means in order to get the tests to come out the same an extra
matrix multiply is needed for most of the CanvasMatrix tests.

Another issue that came up is the Vector Transform test is only
testing 2 libraries. The glMatrix library's vector transform is
mostly bogus. It either does a transform that is only useful in
world space (ie, you can't use it to transform through clipspace)
or it does another one (vec4) that makes absolutely no sense
AFAIK. The short of that is it's vector transform is not
comparible to any other libraries. That leaves only mjs and
tdlMath to be compared. I'm not sure why their results do not
match.

Finally, now that the check is in it shows the the mjs inverse
test is invalid. Arguably mjs should be removed from that test
since it's inverse is limited. None of the other libraries
have the limited form of inverse so there's nothing to compare
it to.