Makes tests use that library's vector format for vectors.
Before most of the tests were using JavaScript arrays for
the vector used in the loops. For example m4.translate was
being passed a JavaScript array.

But arguably those should be passed a vector specific to
that library so I made it do that. Some of the libraries
should show improved perf on FF and Chrome 12/13.