Change timeout mechanism so firefox updates the page.
The old way the tests were all using a bunch of setTimeout of 1ms.
For whatever reason FF never processed the page in that case.

This patch restructures the code so each set of tests is run
and then we wait 100ms to give the browser a chance to show
the results before moving to the next test.