Switch to iterations per second.
The advantage to this way is it's a step on the direction to
not timing out on any browser. Each test is only given
a certain amount of time to run.

The old way had the problem of setting the iteration count
too low meant the timing values were useless for fast machines
and fast browsers. Set it too high and it would timeout on
slow machines or slow browsers.