Use iframes to isolate each library so they don't clash with each other.
Note: Rather than completely refactor the code this version takes
the data from the iframe and then recreates the data that used to be
created by all the calls to testSet inside matrix_benchmark.html

It's probably not the cleanest way to do it but it does mean the
least amount of changes.

Other changes will be needed to make it work in browsers
that don't support TypedArrays.